Added machine learning models and utils for action detection
This update contains the addition of the action detection models 'ADAI' and 'previous models' which are designed for UCF-101 dataset. Components for preprocessing datasets have been built and initialized. Directory utilities are created to manage the file paths. The 'ADAI_train.py' script is also added to train and test the new model. New '.gitignore' file